Don't emit attached add() and remove() for moved items
Regression from 6fbc4b7e7e5aed8739ca1143e0fc1e38b8c8e17a

Reviewed-on: http://codereview.qt.nokia.com/4023

+++ src/declarative/items/qsgitemview.cpp
@@ -1487,8 +1487,10 @@ bool QSGItemViewPrivate::applyModelChanges()
             } else {
                 // removed item
                 removedVisible = true;
-                item->attached->emitRemove();
-                if (item->attached->delayRemove()) {
+                if (!removals[i].isMove())
+                    item->attached->emitRemove();
+
+                if (item->attached->delayRemove() && !removals[i].isMove()) {
                     item->index = -1;
                     QObject::connect(item->attached, SIGNAL(delayRemoveChanged()), q, SLOT(destroyRemoved()), Qt::QueuedConnection);
                     ++it;
